Move sentry feature to sentry.py
