+++ utils.py
@@ -1,14 +1,11 @@
 import os
 from base64 import b64decode, b64encode
-from datetime import datetime
 from time import sleep
 
 import requests
 from Crypto.Cipher import PKCS1_v1_5 as Cipher_PKCS1_v1_5
 from Crypto.PublicKey import RSA
 from dotenv import get_key
-
-from alarm import play_music
 
 # 關掉 pygame 歡迎訊息
 os.environ['PYGAME_HIDE_SUPPORT_PROMPT'] = "hide"
@@ -111,51 +108,3 @@ def get_teammate(user: dict):
         exit(1)
 
     return {'gcid': teammate_gcids, 'nickname': teammate_nicknames}
-
-
-def sentry(teammates: dict):
-    teammate_gcids = teammates['gcid']
-    teammate_nicknames = teammates['nickname']
-
-    cooldown = 3  # 每三分鐘檢查一次, 請善待 PaGamO 伺服器, 不要把他調太低
-    last_seen_land = [0] * len(teammate_gcids)
-
-    print()
-    print(f'開始監視 (每 {cooldown} 分鐘檢查一次, 按 ctrl + c 可停止)')
-
-    while True:
-        try:
-            # 抓隊友各自的領土數
-            current_land = list()
-            for gcid in teammate_gcids:
-                sleep(1)
-                info_resp = s.post(info_url, data={'gc_id': gcid})
-                land_count = info_resp.json()['data']['gamecharacter']['hexagons_count']
-                current_land.append(land_count)
-
-            # 記錄現在時間
-            print(datetime.now().strftime('%m/%d %H:%M'), end=' ')
-
-            attack = False
-            for i in range(len(teammate_gcids)):
-                if current_land[i] < last_seen_land[i]:
-                    attack = True
-                    print('警告: 偵測到入侵!')
-                    print('遭到攻擊的帳號:')
-                    print('{}, 原本土地數: {}, 現在土地數: {}'
-                          .format(teammate_nicknames[i], last_seen_land[i], current_land[i]))
-
-                    play_music()
-
-            if not attack:
-                print('哨兵監視中, 目前土地數: {}'.format(' '.join(str(i) for i in current_land)))
-                sleep(cooldown * 60)  # 請善待 PaGamO 伺服器, 不要把他調太低
-
-            last_seen_land = current_land.copy()
-
-        except Exception as e:
-            print(datetime.now().strftime('%m/%d %H:%M'), end=' ')
-            print(f'發生錯誤: {e}')
-
-            if int(get_key('.env', 'ALERT_IF_SCRIPT_DOWN')):
-                play_music()
